+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.11.0] - 2025-07-21
+### Added
+- Add option in the admin to change the order of address lines
+- Add end-to-end testing
+
+### Changed
+- Change the display of taxes for installment in Argentina
+
 ## [1.10.1] - 2025-06-18
 ### Fixed
 - Correction of discount calculation in CHOPRO processing